Clean up the Custom Formatters sample (#18980)
* Restructure samples folders in view of consistency

* Small changes to the topic content

* Rename .csproj and namespaces for 3.x sample

* Clean up sample and snippets
