--- aspnetcore/web-api/advanced/custom-formatters/sample/Models/Contact.cs
@@ -1,14 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Threading.Tasks;
-
-namespace CustomFormatterDemo.Models
-{
-    public class Contact
-    {
-        public string ID { get; set; }
-        public string FirstName { get; set; }
-        public string LastName { get; set; }
-    }
-}
