+++ aspnetcore/web-api/advanced/custom-formatters/samples/2.x/CustomFormattersSample/2.2sample.csproj
