+++ aspnetcore/web-api/advanced/custom-formatters/samples/2.x/CustomFormattersSample/Models/Contact.cs
