+++ aspnetcore/web-api/advanced/custom-formatters/samples/3.x/CustomFormattersSample/CustomFormattersSample.csproj
@@ -1,8 +1,5 @@
 ﻿<Project Sdk="Microsoft.NET.Sdk.Web">
-
   <PropertyGroup>
     <TargetFramework>netcoreapp3.1</TargetFramework>
   </PropertyGroup>
-
-
 </Project>
