+++ aspnetcore/web-api/advanced/custom-formatters/samples/3.x/CustomFormattersSample/Formatters/VcardInputFormatter.cs
@@ -1,0 +1,90 @@
+﻿using System;
+using System.IO;
+using System.Text;
+using System.Threading.Tasks;
+using CustomFormattersSample.Models;
+using Microsoft.AspNetCore.Mvc.Formatters;
+using Microsoft.Extensions.DependencyInjection;
+using Microsoft.Extensions.Logging;
+using Microsoft.Net.Http.Headers;
+
+namespace CustomFormattersSample.Formatters
+{
+    #region snippet_Class
+    public class VcardInputFormatter : TextInputFormatter
+    {
+        #region snippet_ctor
+        public VcardInputFormatter()
+        {
+            SupportedMediaTypes.Add(MediaTypeHeaderValue.Parse("text/vcard"));
+
+            SupportedEncodings.Add(Encoding.UTF8);
+            SupportedEncodings.Add(Encoding.Unicode);
+        }
+        #endregion
+
+        protected override bool CanReadType(Type type)
+        {
+            return type == typeof(Contact);
+        }
+
+        public override async Task<InputFormatterResult> ReadRequestBodyAsync(
+            InputFormatterContext context, Encoding effectiveEncoding)
+        {
+            var httpContext = context.HttpContext;
+            var serviceProvider = httpContext.RequestServices;
+
+            var logger = serviceProvider.GetRequiredService<ILogger<VcardInputFormatter>>();
+
+            using var reader = new StreamReader(httpContext.Request.Body, effectiveEncoding);
+            string nameLine = null;
+
+            try
+            {
+                await ReadLineAsync("BEGIN:VCARD", reader, context, logger);
+                await ReadLineAsync("VERSION:", reader, context, logger);
+
+                nameLine = await ReadLineAsync("N:", reader, context, logger);
+
+                var split = nameLine.Split(";".ToCharArray());
+                var contact = new Contact
+                {
+                    LastName = split[0].Substring(2),
+                    FirstName = split[1]
+                };
+
+                await ReadLineAsync("FN:", reader, context, logger);
+                await ReadLineAsync("END:VCARD", reader, context, logger);
+
+                logger.LogInformation("nameLine = {nameLine}", nameLine);
+
+                return await InputFormatterResult.SuccessAsync(contact);
+            }
+            catch
+            {
+                logger.LogError("Read failed: nameLine = {nameLine}", nameLine);
+                return await InputFormatterResult.FailureAsync();
+            }
+        }
+
+        private static async Task<string> ReadLineAsync(
+            string expectedText, StreamReader reader, InputFormatterContext context,
+            ILogger logger)
+        {
+            var line = await reader.ReadLineAsync();
+
+            if (!line.StartsWith(expectedText))
+            {
+                var errorMessage = $"Looked for '{expectedText}' and got '{line}'";
+
+                context.ModelState.TryAddModelError(context.ModelName, errorMessage);
+                logger.LogError(errorMessage);
+
+                throw new Exception(errorMessage);
+            }
+
+            return line;
+        }
+    }
+    #endregion
+}
