+++ aspnetcore/web-api/advanced/custom-formatters/samples/3.x/CustomFormattersSample/Models/Contact.cs
@@ -1,0 +1,9 @@
+﻿namespace CustomFormattersSample.Models
+{
+    public class Contact
+    {
+        public string Id { get; set; }
+        public string FirstName { get; set; }
+        public string LastName { get; set; }
+    }
+}
