+++ aspnetcore/web-api/advanced/custom-formatters/samples/3.x/CustomFormattersSample/Startup.cs
@@ -1,23 +1,14 @@
-﻿using CustomFormatterDemo.Formatters;
-using CustomFormatterDemo.Models;
+﻿using CustomFormattersSample.Formatters;
 using Microsoft.AspNetCore.Builder;
 using Microsoft.AspNetCore.Hosting;
-using Microsoft.Extensions.Configuration;
 using Microsoft.Extensions.DependencyInjection;
 using Microsoft.Extensions.Hosting;
 
-namespace CustomFormatterDemo
+namespace CustomFormattersSample
 {
     public class Startup
     {
-        public Startup(IConfiguration configuration)
-        {
-            Configuration = configuration;
-        }
-
-        public IConfiguration Configuration { get; }
-
-        #region mvcoptions
+        #region snippet_ConfigureServices
         public void ConfigureServices(IServiceCollection services)
         {
             services.AddControllers(options =>
