+++ aspnetcore/web-api/advanced/custom-formatters/samples/3.x/CustomFormattersSample/appsettings.Development.json
